Refactor execution into a new class.
